WCM-2909 What is the problem to fix?
    * How to reproduce:
         1. Start WCM 1.2.6
         2. Go to Site Acme -> News
         3. Validate the HTML with W3C validator
            We got validation error :

               1. Warning Line 1610, Column 72: cannot generate system identifier for general entity "workspace"

            ...st/rss/generate?repository=repository&workspace=collaboration&server=http://lo...

               1. Error Line 1610, Column 72: general entity "workspace" not defined and no default entity

            ...st/rss/generate?repository=repository&workspace=collaboration&server=http://lo...

    * Here is the explanation for this error :
An entity reference was found in the document, but there is no reference by that name defined. Often this is caused by misspelling the reference name, unencoded ampersands, or by leaving off the trailing semicolon (. The most common cause of this error is unencoded ampersands in URLs as described by the WDG in "Ampersands in URLs".
Entity references start with an ampersand (&) and end with a semicolon (. If you want to use a literal ampersand in your document you must encode it as "&" (even inside URLs!). Be careful to end entity references with a semicolon or your entity reference may get interpreted in connection with the following text. Also keep in mind that named entity references are case-sensitive; &Aelig; and æ are different characters.
If this error appears in some markup generated by PHP's session handling code, this article has explanations and solutions to your problem.
Note that in most documents, errors related to entity references will trigger up to 5 separate messages from the Validator. Usually these will all disappear when the original problem is fixed.

How is the problem fixed?

    * The problem occurs by existing the URLs in the HTML page which aren't replaced 

      "&" characters to "&amp;"

    * To fix it, found the codes used to setting URL value in the project, add some codes to replace

      "&" characters to "&amp;"

      if it missed