WCM-2925: The selection of repository brings up some content
Problem description
Select the automatic mode in a CLV Portlet, then press the + button to set the repository.
The window to select repositories appears.

In a directory, all sub-directories can be selected, but it is also possible to select some types of content (like the node type exo:link). These nodes type have the supertype nt:unstructured.

Fix description
* Nodes are not well filtered when the purpose is only to show folder path to choose hence some node types are shown unwantedly and cause the problem.
* Create function filters (isDocumentType) in UpdateTreeNodeConnecter.java and UISelectPathPanelFolder.java to filter out the unwanted nodes.